finished network refactor and debugging webrtc connection

+++ uno/src/network/protocol/hand_size.rs
@@ -6,12 +6,12 @@ use naia_shared::{Property, Replicate};
 #[derive(Component, Replicate)]
 #[protocol_path = "crate::network::protocol::Protocol"]
 pub struct HandSize {
-    pub size: Property<u8>,
+    pub size: Property<usize>,
     pub player_id: Property<String>,
 }
 
 impl HandSize {
-    pub fn new(size: u8, player_id: Uuid) -> Self {
+    pub fn new(size: usize, player_id: Uuid) -> Self {
         HandSize::new_complete(size, player_id.to_string())
     }
 }
